Ajout de donnee supplementaires dans la premiere version du simulateur

+++ src/main/java/data/WallPart.java
@@ -29,7 +29,7 @@ public class WallPart {
         return compos;
     }
 
-    public void setTemp(int newTemp){
+    public void setTemp(double newTemp){
         this.temp=newTemp;
     }
 }
